vitpose/requirements.txt
ffmpeg==1.4
matplotlib==3.6.2
munkres==1.1.4
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
easydict